implemented wriyter list to avoid vector

+++ src/DLogLevelMap.cpp
@@ -44,7 +44,7 @@ void DLogLevelMap::setLevel(const char* tag, DLogLevel level)
 
 DLogLevelEntry* DLogLevelMap::find(const char* tag)
 {
-    for(int i = 0; i < _level_entry_count; ++i)
+    for(size_t i = 0; i < _level_entry_count; ++i)
     {
         if (strcmp(tag, _level_entries[i].tag) == 0)
         {
